Don't release code in kolibri-tools that isn't needed outside the kolibri repo.
